1. 更改ECPM类型

+++ adsdk/src/main/java/com/hauyu/adsdk/adloader/adfb/FBLoader.java
@@ -600,7 +600,7 @@ public class FBLoader extends AbstractSdkLoader {
                 if (getAdListener() != null) {
                     AdReward adReward = new AdReward();
                     adReward.setType(Constant.ECPM);
-                    int ecpm = 0;
+                    double ecpm = 0;
                     if (mPidConfig != null) {
                         ecpm = mPidConfig.getEcpm();
                     }
