+++ adsdk/src/main/java/com/hauyu/adsdk/adloader/base/AbstractSdkLoader.java
@@ -576,7 +576,7 @@ public abstract class AbstractSdkLoader implements ISdkLoader, Handler.Callback 
     }
 
     @Override
-    public int getEcpm() {
+    public double getEcpm() {
         if (mPidConfig != null) {
             return mPidConfig.getEcpm();
         }
